Remove nokogiri dependency and use REXML for XML parsing.
This removes nokogiri as a dependency in favor of Ruby's built-in REXML
parser. Since nokogiri can be a somewhat large dependency that needs to
be compiled, this removes that as a hard dependency for using
omniauth-cas.

This changes the documented response of `user_info` being nil for
invalid responses, but the tests seem to contradict this, since it was
always returning an empty hash for failure responses. This stems from
the fact that I don't think `find_authentication_success` would ever
return nil because Nokogiri::XML::XPath::SyntaxError would never be thrown
(that would only be thrown if the xpath expression was invalid). So even
if the document didn't match the XPath expression, Nokogiri always
returned an empty result, rather than nil. So functionally, I don't
think anything's changed, it just fixes the documentation of
`user_info`.